Additions to output boundary for account and user
