+++ src/main/java/Boundaries/InputBoundary.java
@@ -1,4 +1,7 @@
 package Boundaries;
+import ResponseModels.UserAccountResponseModel;
+import RequestModels.UserAccountRequestModel;
 
 public interface InputBoundary {
+    UserAccountResponseModel create(UserAccountRequestModel requestModel);
 }
